Move most of io updating to common.rs
